refactor(#2744): re-home ModelCapability into backend_model.rs — anti-monolith ceiling
Coverage's src_file_size_invariant caught backend.rs at 2637 LOC
(ceiling 2500) after the PR-A additions. Cohesive split per the
invariant's guidance: ModelCapability/ModelFlagHit + scan + the
per-backend capability table + their tests move to src/backend_model.rs
(207 LOC); backend.rs keeps a delegating Backend::model_capability and
re-exports the types, so every crate::backend:: path is unchanged.
backend.rs is back to 2451 LOC. No behavior change — 2744 suite 22/22
and the invariant now pass locally.

Refs #2744


Agend-Agent: claude-aef7c0
Agend-Task: t-20260712101738772211-40783-67
Agend-Branch: feat/model-persistence-explicit-intent
Agend-Issued-At: 2026-07-12T10:18:11.812063+00:00
