+++ src/main.rs
@@ -16,6 +16,7 @@ mod app;
 mod auth_cookie;
 mod backend;
 mod backend_harness;
+mod backend_model;
 mod backend_profile;
 mod behavioral;
 mod binding;
